refactor: gam scripts and model
