+++ includes/class-blocks.php
@@ -8,10 +8,7 @@
 namespace Newspack_Ads;
 
 use Newspack_Ads\Core;
-use Newspack_Ads\Settings;
-use Newspack_Ads\Providers;
 use Newspack_Ads\Placements;
-use Newspack_Ads\Model;
 
 /**
  * Newspack Ads Blocks Management
@@ -26,8 +23,6 @@ final class Blocks {
 	public static function init() {
 		add_action( 'init', [ __CLASS__, 'register_block' ] );
 		add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'enqueue_block_editor_assets' ] );
-		add_action( 'wp_head', [ __CLASS__, 'insert_gpt_header_script' ] );
-		add_action( 'wp_footer', [ __CLASS__, 'insert_gpt_footer_script' ] );
 	}
 
 	/**
@@ -202,213 +197,6 @@ final class Blocks {
 			array_push( $classes, sprintf( 'has-%s-background-color', $attributes['backgroundColor'] ) );
 		}
 		return implode( ' ', $classes );
-	}
-
-	/**
-	 * Google Publisher Tag header script.
-	 */
-	public static function insert_gpt_header_script() {
-		if ( ! newspack_ads_should_show_ads() ) {
-			return;
-		}
-		if ( ! Providers::is_provider_active( 'gam' ) ) {
-			return;
-		}
-		if ( Core::is_amp() ) {
-			return;
-		}
-		?>
-		<script async src="https://securepubads.g.doubleclick.net/tag/js/gpt.js" data-amp-plus-allowed></script>
-		<?php
-	}
-
-	/**
-	 * Google Publisher Tag configuration script.
-	 */
-	public static function insert_gpt_footer_script() {
-		if ( ! newspack_ads_should_show_ads() ) {
-			return;
-		}
-		if ( ! Providers::is_provider_active( 'gam' ) ) {
-			return;
-		}
-		if ( Core::is_amp() ) {
-			return;
-		}
-
-		$network_code = Model::get_active_network_code();
-
-		$prepared_unit_data = [];
-		foreach ( Model::$ad_ids as $unique_id => $ad_unit ) {
-			$ad_targeting = Model::get_ad_targeting( $ad_unit );
-
-			$container_id = esc_attr( 'div-gpt-ad-' . $unique_id . '-0' );
-			$sizes        = $ad_unit['sizes'];
-
-			if ( ! is_array( $sizes ) ) {
-				$sizes = [];
-			}
-
-			// Remove all ad sizes greater than 600px wide for sticky ads.
-			if ( Model::is_sticky( $ad_unit ) ) {
-				$sizes = array_filter(
-					$sizes,
-					function( $size ) {
-						return $size[0] < 600;
-					}
-				);
-				$sizes = array_values( $sizes );
-			}
-
-			$prepared_unit_data[ $container_id ] = [
-				'unique_id' => $unique_id,
-				'name'      => esc_attr( $ad_unit['name'] ),
-				'code'      => esc_attr( $ad_unit['code'] ),
-				'sizes'     => $sizes,
-				'fluid'     => (bool) $ad_unit['fluid'],
-				'targeting' => $ad_targeting,
-				'sticky'    => Model::is_sticky( $ad_unit ),
-				'size_map'  => Model::get_responsive_size_map( $sizes ),
-			];
-		}
-
-		$ad_config = [
-			'network_code'         => esc_attr( $network_code ),
-			'disable_initial_load' => (bool) apply_filters( 'newspack_ads_disable_gtag_initial_load', false ),
-		];
-	
-		/**
-		 * Filters the ads data parsed for gtag.
-		 *
-		 * @param array $data {
-		 *   Ads data parsed for gtag inline script.
-		 *
-		 *   @type string $unique_id Unique ID for the ad unit.
-		 *   @type string $name      Ad name.
-		 *   @type string $code      Ad code.
-		 *   @type array  $sizes     Ad sizes.
-		 *   @type bool   $fluid     Whether the ad is fluid.
-		 *   @type array  $targeting Ad targeting.
-		 *   @type bool   $sticky    Whether the ad is sticky.
-		 * }
-		 */
-		$prepared_unit_data = apply_filters( 'newspack_ads_gtag_ads_data', $prepared_unit_data );
-
-		do_action( 'newspack_ads_gtag_before_script', $ad_config, $prepared_unit_data );
-		?>
-		<script data-amp-plus-allowed>
-			window.googletag = window.googletag || { cmd: [] };
-			googletag.cmd.push(function() {
-				var ad_config        = <?php echo wp_json_encode( $ad_config ); ?>;
-				var all_ad_units     = <?php echo wp_json_encode( $prepared_unit_data ); ?>;
-				var lazy_load        = <?php echo wp_json_encode( Settings::get_settings( 'lazy_load', true ), JSON_FORCE_OBJECT ); ?>;
-				var defined_ad_units = {};
-
-				for ( var container_id in all_ad_units ) {
-					var ad_unit = all_ad_units[ container_id ];
-
-					// Only set up ad units that are present on the page.
-					if ( ! document.querySelector( '#' + container_id ) ) {
-						continue;
-					}
-
-					var slotSizes = ad_unit['sizes'];
-					if ( ad_unit['fluid'] ) {
-						slotSizes = slotSizes.concat( 'fluid' );
-					}
-
-					defined_ad_units[ container_id ] = googletag.defineSlot(
-						'/' + ad_config['network_code'] + '/' + ad_unit['code'],
-						slotSizes,
-						container_id
-					).addService( googletag.pubads() );
-
-					for ( var target_key in ad_unit['targeting'] ) {
-						defined_ad_units[ container_id ].setTargeting( target_key, ad_unit['targeting'][ target_key ] );
-					}
-
-					/**
-					 * Configure responsive ads.
-					 * Ads wider than the viewport should not show.
-					 */
-
-					// Get all of the unique ad widths.
-					var unique_widths = {};
-					ad_unit['sizes'].forEach( function( size ) {
-						unique_widths[ size[0] ] = [];
-					} );
-
-					// For each width, get all of the sizes equal-to-or-smaller than it.
-					for ( width in unique_widths ) {
-						ad_unit['sizes'].forEach( function( size ) {
-							if ( size[0] <= width ) {
-								unique_widths[ width ].push( size );
-							}
-						} );
-					}
-
-					// Build and set the responsive mapping.
-					// @see https://developers.google.com/doubleclick-gpt/guides/ad-sizes#responsive_ads
-					var mapping = googletag.sizeMapping();
-					// Default base is to not show ads.
-					var baseSizes = [];
-					// If the ad unit is fluid, base includes fluid.
-					if( ad_unit['fluid'] ) {
-						baseSizes = baseSizes.concat( 'fluid' );
-					}
-					// Iterate through size map.
-					for ( viewportWidth in ad_unit['size_map'] ) {
-						var mappedSizes = ad_unit['size_map'][ viewportWidth ];
-						mapping.addSize( [ parseInt( viewportWidth ), 0 ], baseSizes.concat( mappedSizes ) );
-					}
-					// Sticky ads should only be shown on mobile (screen width <=600px).
-					if ( ad_unit['sticky'] ) {
-						mapping.addSize( [600, 0], baseSizes );
-					}
-					// On viewports smaller than the smallest ad size, don't show any ads.
-					mapping.addSize( [0, 0], baseSizes );
-					defined_ad_units[ container_id ].defineSizeMapping( mapping.build() );
-				}
-
-				if ( ad_config['disable_initial_load'] ) {
-					googletag.pubads().disableInitialLoad();
-				}
-				googletag.pubads().collapseEmptyDivs();
-				googletag.pubads().enableSingleRequest();
-				if ( lazy_load && lazy_load.active ) {
-					googletag.pubads().enableLazyLoad( {
-						fetchMarginPercent: lazy_load.fetch_margin_percent,
-						renderMarginPercent: lazy_load.render_margin_percent,
-						mobileScaling: lazy_load.mobile_scaling
-					} );
-				}
-				googletag.enableServices();
-
-				for ( var container_id in defined_ad_units ) {
-					googletag.display( container_id );
-				}
-				// Identify fluid rendered ad and fix iframe width.
-				// GPT currently sets the iframe with `min-width` set to 100% and property `width` set to 0.
-				// This causes the iframe to be rendered with 0 width.
-				googletag.pubads().addEventListener( 'slotRenderEnded', function(event) {
-					var sizes = event.slot.getSizes();
-					if (
-						( event.size === null || event.size[0] === 0 ) &&
-						Array.isArray( sizes ) && sizes.indexOf( 'fluid' ) !== -1
-					) {
-						var container = document.getElementById( event.slot.getSlotElementId() );
-						if ( container ) {
-							var iframe = container.querySelector( 'iframe' );
-							if ( iframe ) {
-								iframe.style.width = '100%';
-							}
-						}
-					}
-				} );
-			} );
-		</script>
-		<?php
-		do_action( 'newspack_ads_gtag_after_script', $ad_config, $prepared_unit_data );
 	}
 }
 Blocks::init();
